backend/requirements.txt
Flask==3.1.1
Flask-CORS==6.0.0
requests==2.32.3
python-dotenv==1.1.1
gunicorn==21.2.0

